Cleanup and small enhancements
