+++ src/HaKafkaNet/Implementations/Automations/Prebuilt/LightOffOnNoMotion.cs
@@ -1,4 +1,6 @@
 ﻿
+using System.Text.Json;
+
 namespace HaKafkaNet;
 
 [ExcludeFromDiscovery]
@@ -6,15 +8,13 @@ public class LightOffOnNoMotion : ConditionalAutomationBase
 {
     private readonly List<string> _motionIds = new();
     private readonly List<string> _lightIds = new();
-    private readonly TimeSpan _duration;
     private readonly IHaServices _services;
 
     public LightOffOnNoMotion(IEnumerable<string> motionIds, IEnumerable<string> lightIds, TimeSpan duration, IHaServices services)
-        : base(motionIds, duration)
+        : base(motionIds.Union(lightIds), duration)
     {
         this._motionIds.AddRange(motionIds);
         this._lightIds.AddRange(lightIds);
-        this._duration = duration;
         this._services = services;
     }
 
@@ -23,19 +23,19 @@ public class LightOffOnNoMotion : ConditionalAutomationBase
     {
         var motionStates = 
             from m in _motionIds
-            select _services.EntityProvider.GetEntity(m, cancellationToken); // all should be off
+            select _services.EntityProvider.GetOnOffEntity(m, cancellationToken); // all should be off
         
         var lightStates = 
             from l in _lightIds
-            select _services.EntityProvider.GetEntity(l, cancellationToken);  // any should be on
+            select _services.EntityProvider.GetOnOffEntity(l, cancellationToken);  // any should be on
 
-        Task<HaEntityState[]> motionResults;
-        Task<HaEntityState[]> lightResults;
+        Task<HaEntityState<OnOff, JsonElement>?[]> motionResults;
+        Task<HaEntityState<OnOff, JsonElement>?[]> lightResults;
 
         return Task.WhenAll(
             motionResults = Task.WhenAll(motionStates),
             lightResults = Task.WhenAll(lightStates)).ContinueWith(t =>
-                motionResults.Result.All(m => m.State == "off") && lightResults.Result.Any(l => l.State == "on")
+                motionResults.Result.All(m => m?.State == OnOff.Off) && lightResults.Result.Any(l => l?.State == OnOff.On)
             ,cancellationToken, TaskContinuationOptions.NotOnFaulted, TaskScheduler.Current);        
     }
 
@@ -45,7 +45,7 @@ public class LightOffOnNoMotion : ConditionalAutomationBase
             from lightId in _lightIds
             select _services.EntityProvider.GetOnOffEntity(lightId, cancellationToken)
                 .ContinueWith(t => 
-                    t.Result?.State == OnOff.On
+                    t.Result?.Bad() != true && t.Result?.State == OnOff.On
                         ? _services.Api.TurnOff(lightId, cancellationToken)
                         : Task.CompletedTask
                 , cancellationToken, TaskContinuationOptions.NotOnFaulted, TaskScheduler.Current)
