+++ src/HaKafkaNet/Implementations/Automations/Prebuilt/SunAutomations.cs
@@ -9,7 +9,8 @@ public abstract class SunAutomation : SchedulableAutomationBase
         EventTiming.PostStartup | EventTiming.PreStartupSameAsLastCached | 
         EventTiming.PreStartupPostLastCached | EventTiming.PreStartupNotCached;
 
-    public SunAutomation(Func<CancellationToken, Task> execution, TimeSpan? offset = null, EventTiming timings = DEFAULT_SUN_EVENT_TIMINGS, bool executePast = true): base(["sun.sun"], executePast)
+    public SunAutomation(Func<CancellationToken, Task> execution, TimeSpan? offset = null, EventTiming timings = DEFAULT_SUN_EVENT_TIMINGS, bool executePast = true)
+        : base(["sun.sun"], executePast)
     {
         base.IsReschedulable = false;
         base.EventTimings = timings;
