+++ src/HaKafkaNet/Models/EntityModels/HaEntityStateChange.cs
@@ -44,9 +44,31 @@ public static class StateChangeExtensions
         };
     }    
 
+    /// <summary>
+    /// Use when you want to specify both state and attribute types
+    /// </summary>
+    /// <typeparam name="Tstate">The State Type</typeparam>
+    /// <typeparam name="Tatt">The Type for Attributes</typeparam>
+    /// <param name="change"></param>
+    /// <returns></returns>
     public static HaEntityStateChange<HaEntityState<Tstate,Tatt>> ToTyped<Tstate, Tatt>(this HaEntityStateChange change) =>  Transform<HaEntityState<Tstate,Tatt>, Tstate, Tatt>(change);
+
+    /// <summary>
+    /// Use when you want to only specify the Ttate type. Attributes will be JsonElement
+    /// </summary>
+    /// <typeparam name="Tstate">The State Type</typeparam>
+    /// <param name="change"></param>
+    /// <returns></returns>
     public static HaEntityStateChange<HaEntityState<Tstate,JsonElement>> ToStateTyped<Tstate>(this HaEntityStateChange change) =>  Transform<HaEntityState<Tstate,JsonElement>, Tstate, JsonElement>(change);
+
+    /// <summary>
+    /// Use when you want to only specify the Attributes type. State will be string
+    /// </summary>
+    /// <typeparam name="Tatt">The Type for Attributes</typeparam>
+    /// <param name="change"></param>
+    /// <returns></returns>
     public static HaEntityStateChange<HaEntityState<string,Tatt>> ToAttributeTyped<Tatt>(this HaEntityStateChange change) =>  Transform<HaEntityState<string,Tatt>, string, Tatt>(change);
+
     public static HaEntityStateChange<HaEntityState<OnOff, JsonElement>> ToOnOff(this HaEntityStateChange change) => Transform<HaEntityState<OnOff, JsonElement>, OnOff, JsonElement>(change);
     public static HaEntityStateChange<HaEntityState<OnOff, T>> ToOnOff<T>(this HaEntityStateChange change) =>  Transform<HaEntityState<OnOff, T>, OnOff, T>(change);
     public static HaEntityStateChange<HaEntityState<int?, JsonElement>> ToIntTyped(this HaEntityStateChange change) =>  Transform<HaEntityState<int?, JsonElement>, int?, JsonElement>(change);
@@ -57,7 +79,8 @@ public static class StateChangeExtensions
     public static HaEntityStateChange<HaEntityState<DateTime?, T>> ToDateTimeTyped<T>(this HaEntityStateChange change) =>  Transform<HaEntityState<DateTime?, T>, DateTime?, T>(change);
     public static HaEntityStateChange<HaEntityState<BatteryState, JsonElement>> ToBatteryState(this HaEntityStateChange change) =>  Transform<HaEntityState<BatteryState, JsonElement>, BatteryState, JsonElement>(change);
     public static HaEntityStateChange<HaEntityState<BatteryState, T>> ToBatteryState<T>(this HaEntityStateChange change) =>  Transform<HaEntityState<BatteryState, T>, BatteryState, T>(change);
+    public static HaEntityStateChange<HaEntityState<SunState, SunAttributes>> ToSun(this HaEntityStateChange change) =>  Transform<HaEntityState<SunState, SunAttributes>, SunState, SunAttributes>(change);
+    public static HaEntityStateChange<HaEntityState<OnOff, LightModel>> ToLight(this HaEntityStateChange change) => Transform<HaEntityState<OnOff, LightModel>, OnOff, LightModel>(change);
+    public static HaEntityStateChange<HaEntityState<OnOff, ColoredLightModel>> ToColoredLight(this HaEntityStateChange change) => Transform<HaEntityState<OnOff, ColoredLightModel>, OnOff, ColoredLightModel>(change);
     public static HaEntityStateChange<HaEntityState<DateTime?, SceneControllerEvent>> ToSceneControllerEvent(this HaEntityStateChange change) => Transform<HaEntityState<DateTime?, SceneControllerEvent>, DateTime?, SceneControllerEvent>(change);
-    
-    //public static HaEntityStateChange<SunModel> ToSun(this HaEntityStateChange change) =>  change.Transform<SunModel, SunState, SunAttributes>();
 }
